+++ support-frontend/assets/components/checkoutForm/checkoutForm.jsx
@@ -1,16 +1,10 @@
 // @flow
-// eslint-disable-next-line no-unused-vars
+
 import React, { type Node, type Element, type ChildrenArray } from 'react';
-/** @jsx jsx */ import { jsx, css } from '@emotion/core';
-import { headline } from '@guardian/src-foundations/typography';
 import { type Option } from 'helpers/types/option';
 import Heading, { type HeadingSize } from 'components/heading/heading';
 
 import './checkoutForm.scss';
-
-const h2 = css`
-    ${headline.small()};
-`;
 
 /*
 Form Section
@@ -43,33 +37,6 @@ FormSection.defaultProps = {
   id: '',
 };
 
-// Hidden version of form section
-
-type FormSectionHiddenPropTypes = {|
-  title: Option<string>,
-  children: Node,
-  show?: boolean,
-  id?: Option<string>,
-|};
-
-const FormSectionHiddenUntilSelected = ({
-  children, title, show, id,
-}: FormSectionHiddenPropTypes) => (
-  <div id={id} className={show ? 'component-checkout-form-section' : 'component-checkout-form-section is-hidden'}>
-    {show && (
-    <div className="component-checkout-form-section__wrap">
-      {title && <Heading css={h2}>{title}</Heading>}
-      {children}
-    </div>)}
-  </div>
-);
-
-FormSectionHiddenUntilSelected.defaultProps = {
-  title: null,
-  show: false,
-  id: '',
-};
-
 /*
 Form
 the top level form itself
@@ -81,4 +48,4 @@ type FormPropTypes = {
 const Form = ({ children, ...otherProps }: FormPropTypes) => (<form {...otherProps} className="component-checkout-form">{children}</form>);
 
 export default Form;
-export { FormSection, FormSectionHiddenUntilSelected };
+export { FormSection };
