+++ support-frontend/assets/components/checkoutForm/checkoutForm.scss
@@ -1,7 +1,7 @@
 @import '~stylesheets/gu-sass/gu-sass';
 
 .component-checkout-form-section__wrap {
-    max-width: gu-span(5)
+    max-width: gu-span(5);
 }
 
 .component-checkout-form-section {
